Struct presort::permuted_vec::PermutedVec

pub struct PermutedVec<T> {
    // some fields omitted
}

The type of permuted vectors.

Methods

impl<T> PermutedVec<T>

fn new() -> PermutedVec<T>

Create a new, empty presorted vector.

fn permutation_iter(&self) -> Iter<usize>

An iterator over the permutation

fn permuted_iter(&self) -> PermutedIter<T>

An iterator over the permuted vector

fn is_sorted_by<F>(&self, f: &mut F) -> bool where F: FnMut(&T, &T) -> Ordering

Is the permuted vector already sorted?

fn sort_by<F>(&mut self, f: F) where F: FnMut(&T, &T) -> Ordering

Sort the permutation on the vector

fn sorted_iter_by<F>(&mut self, f: F) -> PermutedIter<T> where F: FnMut(&T, &T) -> Ordering

A sorted iterator over the vector.

fn get(&self, index: usize) -> Option<&T>

Get the ith element of the vector. Returns None if the vector contains fewer than i elements.

fn get_permuted(&self, permuted: usize) -> Option<&T>

Get the ith element of the permuted vector. Returns None if the vector contains fewer than i elements.

fn set(&mut self, index: usize, value: T)

Set the ith element of the vector. Panics if the vector contains fewer than i elements.

fn push(&mut self, value: T)

Append an element to the end of the vector.

fn truncate(&mut self, len: usize)

Truncate this vector and reset the sort if necessary.

fn len(&self) -> usize

The length of the vector.